MAINT: fix partially build on windows.
