+++ cmake/UseDOSaraGraphics.cmake
@@ -1,15 +1,9 @@
-find_package(Qt5 COMPONENTS Core Widgets OpenGL REQUIRED)
-find_package(OpenGL REQUIRED)
-
-add_definitions(${Qt5Widgets_DEFINITIONS})
-if (UNIX)
-  set(CMAKE_CXX_FLAGS
-      "${CMAKE_CXX_FLAGS} ${Qt5Widgets_EXECUTABLE_COMPILE_FLAGS}")
-endif ()
-
 if (SARA_USE_FROM_SOURCE)
   get_property(DO_Sara_Graphics_ADDED GLOBAL PROPERTY _DO_Sara_Graphics_INCLUDED)
   if (NOT DO_Sara_Graphics_ADDED)
+    find_package(Qt5 COMPONENTS Core Widgets OpenGL REQUIRED)
+    find_package(OpenGL REQUIRED)
+
     set(CMAKE_AUTOMOC ON)
     set(CMAKE_INCLUDE_CURRENT_DIR ON)
 
@@ -19,21 +13,19 @@ if (SARA_USE_FROM_SOURCE)
 
     target_include_directories(DO_Sara_Graphics
       PUBLIC
-      ${Qt5Widgets_INCLUDE_DIRS}
-      ${Qt5OpenGL_INCLUDE_DIRS}
       ${DO_Sara_INCLUDE_DIR})
+
     target_link_libraries(DO_Sara_Graphics
       PUBLIC
-      Qt5::Widgets Qt5::OpenGL
+      Qt5::Widgets
+      Qt5::OpenGL
       ${OPENGL_LIBRARIES})
-    target_compile_definitions(DO_Sara_Graphics
-      PRIVATE
-      -DGL_SILENCE_DEPRECATION)
 
     if (WIN32)
       target_link_libraries(DO_Sara_Graphics PRIVATE Qt5::WinMain)
     endif ()
 
+    target_compile_definitions(DO_Sara_Graphics PRIVATE GL_SILENCE_DEPRECATION)
 
     set(CMAKE_AUTOMOC OFF)
     set(CMAKE_INCLUDE_CURRENT_DIR OFF)
