+++ cpp/src/DO/Sara/Features/IO.hpp
@@ -28,21 +28,53 @@
 #include <vector>
 
 
-
 namespace DO { namespace Sara {
 
   /*!
-    @addtogroup Features
-    @{
-  */
+   *  @addtogroup Features
+   *  @{
+   */
+
+  template <>
+  struct CalculateH5Type<OERegion::Type>
+  {
+    static inline auto value()
+    {
+      return H5::PredType::NATIVE_UINT8;
+    };
+  };
+
+  template <>
+  struct CalculateH5Type<OERegion::ExtremumType>
+  {
+    static inline auto value()
+    {
+      return H5::PredType::NATIVE_INT8;
+    };
+  };
+
+  template <>
+  struct CalculateH5Type<OERegion>
+  {
+    static inline auto value() -> H5::CompType
+    {
+      auto h5_comp_type = H5::CompType{sizeof(OERegion)};
+      INSERT_MEMBER(h5_comp_type, OERegion, coords);
+      INSERT_MEMBER(h5_comp_type, OERegion, shape_matrix);
+      INSERT_MEMBER(h5_comp_type, OERegion, orientation);
+      INSERT_MEMBER(h5_comp_type, OERegion, extremum_value);
+      INSERT_MEMBER(h5_comp_type, OERegion, type);
+      INSERT_MEMBER(h5_comp_type, OERegion, extremum_type);
+      return h5_comp_type;
+    }
+  };
 
   template <typename T>
   bool read_keypoints(std::vector<OERegion>& features,
-                      Tensor_<T, 2>& descriptors,
-                      const std::string& name)
+                      Tensor_<T, 2>& descriptors, const std::string& name)
   {
     using namespace std;
-    ifstream file{ name.c_str() };
+    ifstream file{name.c_str()};
     if (!file.is_open())
     {
       cerr << "Can't open file " << name << endl;
@@ -85,7 +117,7 @@ namespace DO { namespace Sara {
     }
 
     file << features.size() << " " << descriptors.cols() << std::endl;
-    for(size_t i = 0; i < features.size(); ++i)
+    for (size_t i = 0; i < features.size(); ++i)
     {
       const OERegion& feat = features[i];
 
@@ -94,9 +126,10 @@ namespace DO { namespace Sara {
       file << feat.orientation << ' ';
       file << int(feat.type) << ' ';
 
-      file << Map<const Matrix<T, 1, Dynamic> >(
-        descriptors[static_cast<int>(i)].data(),
-        1, descriptors.cols() ) << endl;
+      file << Map<const Matrix<T, 1, Dynamic>>(
+                  descriptors[static_cast<int>(i)].data(), 1,
+                  descriptors.cols())
+           << endl;
     }
 
     file.close();
@@ -128,5 +161,4 @@ namespace DO { namespace Sara {
 
   //! @}
 
-} /* namespace Sara */
-} /* namespace DO */
+}}  // namespace DO::Sara
